Integrate easy-connect, fix serial, fix warning
Integrate easy-connect
- we can remove all the external RF-drivers from main folder
- get easy-connect into use
- remove all the complicated network stuff

Fix serial - we don't need raw serial and we don't need to initialize it
if we set the right stuff to mbed_app.json (added). Plain printf() works
and it even does type checking now. (Points to Kevin Bracey on this).

Start the thread in a non-deprecated way using callback to remove one compiler
warning (again, points to Kevin Bracey).

Easy-connect ESP8266 TX/RX pins in different defines.
mbed-os-example-client has NEVER used 2 sets of WiFi pins, so need
to add a define to hack them into one set - otherwise it will not compile.

The cp command would not have worked if one would have changed
the compiler. Add UBLOX ODIN ETH config to be built as well,
now that we have figured out how to make the the ethernet work
on it.

Fix also the WIFI (or make it worse, depending on your opinion) to
work with easy-connect. Easy-connect has 2 separate WIFIs, so you
have to choose between ESP8266 vs. ODIN. Thus, 2 config files
are also needed, one for ESP8266 and one for ODIN.