Using thread signals instead of restarting a thread
The blinky thread was being restarted everytime a POST was received.
This wasn't working correctly for multiple POST requests, it would only
work for the first POST request. This is also a very heavyweight action.
This uses thread signals instead and keeps the blink thread alive for
the whole application